+++ __fixtures__/copy/src/shared/i18n/locales/zh-CN/common.ts
@@ -1,0 +1,4 @@
+export default {
+  save: '保存',
+  orphan: '没人用',
+}