+++ __fixtures__/copy/src/shared/i18n/locales/zh-CN/index.ts
@@ -1,0 +1,4 @@
+import common from './common'
+import nav from './nav'
+
+export default { common, nav }